refactor: socket 모듈 위치 변경 및 import 경로 수정
