+++ src/main.ts
@@ -1,6 +1,6 @@
 import { createServer, startServer } from './server-utils.js';
 import { configureServer, registerPlugins, setupGracefulShutdown } from './server-config.js';
-import { createSocketServer } from './socket.js';
+import { createSocketServer } from './plugins/socket.js';
 
 async function init() {
   const server = createServer();
